delete sqlite,fixbug delete,merge dialog logout and delete todo
