+++ lib/l10n/app_en.arb
@@ -24,8 +24,8 @@
  "logOut": "Log out",
  "editTodo": "Edit todo",
  "doYouWant": "Do you want to ",
-  "yes": "Yes (Y)",
- "no": "No (N)",
+  "yes": "Yes",
+ "no": "No",
  "synchronizedTxt": "Some data is not synchronized",
  "delete": "delete",
  "isEmpty": "This field cannot be left blank",
