+++ lib/l10n/app_vi.arb
@@ -9,7 +9,7 @@
   "noData": "Không có dữ liệu",
   "myScores": "Điểm của tôi",
   "featureSoon": "Tính năng này sẽ sớm ra mắt",
-  "createTodo": "Tạo việc làm",
+  "createTodo": "Tạo ghi chú",
   "title": "Tiêu đề",
   "setDate": "Đặt ngày",
   "setTime":"Đặt giờ",
@@ -22,10 +22,10 @@
   "aboutUs": "Liên hệ",
   "rateMe": "Đánh giá",
   "logOut": "Đăng xuất",
-  "editTodo": "Chỉnh sửa việc làm",
+  "editTodo": "Sửa ghi chú",
   "doYouWant": "Bạn có muốn ",
-  "yes": "Có (C)",
-  "no": "Không (K)",
+  "yes": "Có",
+  "no": "Không",
   "synchronizedTxt": "Một số dữ liệu không được đồng bộ hóa",
   "delete": "xóa",
   "isEmpty": "Không được phép bỏ trống ô này",
