+++ lib/main.dart
@@ -6,16 +6,14 @@ import 'package:schedule/common/config/local_config.dart';
 import 'package:schedule/common/constants/key_constants.dart';
 import 'package:schedule/common/injector/injector.dart';
 import 'package:schedule/presentation/%20language_select/%20language_select.dart';
-import 'package:schedule/service/database/database.dart';
+
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'presentation/app.dart';
 
 Future main() async {
   WidgetsFlutterBinding.ensureInitialized();
-
   ///*BlocSupervisor.delegate =*/ SimpleBlocDelegate();
-  await DatabaseCreator().initDatabase();
   Injector.setup();
   final hiveSetUp = Injector.getIt<LocalConfig>();
   await hiveSetUp.init();
