--- lib/service/providers/providers.dart
@@ -1,2 +1,0 @@
-export './online/provider_online.dart';
-export 'offline/provider_offline.dart';
