--- lib/service/repositors/online/repository_online.dart
@@ -1,7 +1,0 @@
-import 'package:schedule/service/services.dart';
-
-class RepositoryOnline {
-  final ProviderOnline _provider = ProviderOnline();
-  Future<String> fetchScheduleSchoolDataRepo(String account, String password) =>
-      _provider.fetchScheduleSchoolDataProvider(account, password);
-}
