--- lib/service/repositors/repositories.dart
@@ -1,2 +1,0 @@
-export 'offline/repository_offline.dart';
-export 'online/repository_online.dart';
