+++ lib/service/services.dart
@@ -1,5 +1,3 @@
-export './database/database.dart';
-export 'providers/providers.dart';
-export 'repositors/repositories.dart';
+
 export 'share_service.dart';
 export 'un_authorized_exception.dart';
